Добавила основную логику деталей займа

+++ app/src/main/java/com/nsu/focusstartproject/data/loans_network/LoanApi.kt
@@ -12,7 +12,7 @@ import retrofit2.http.Path
 interface LoanApi {
 
     @GET("/loans/{id}")
-    suspend fun getLoan(@Path("id") id: Int): Response<Loan>
+    suspend fun getLoan(@Path("id") id: Long): Response<Loan>
 
     @GET("/loans/all")
     suspend fun getAllLoans(): Response<List<Loan>>
