+++ app/src/main/java/com/nsu/focusstartproject/presentation/auth_user_screens/loan_list/LoanListAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.nsu.focusstartproject.domain.Loan
 
 class LoanListAdapter(
-    private val onClick: () -> Unit
+    private val onClick: (Long) -> Unit
 ) : RecyclerView.Adapter<LoanHolder>() {
 
     var loans: List<Loan> = emptyList()
